feat: implement permission checks for B3 Storage and IPAL controllers, update related tests

+++ app/Http/Controllers/Api/B3Storage/B3StorageInitiatorDepartmentController.php
@@ -6,11 +6,15 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\B3Storage\SaveB3StorageInitiatorDepartmentRequest;
 use App\Models\B3Storage\B3StorageInitiatorDepartment;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 
 class B3StorageInitiatorDepartmentController extends Controller
 {
-    public function index(): JsonResponse
+    public function index(Request $request): JsonResponse
     {
+        abort_unless($request->user()?->can('b3storage.master.view'), Response::HTTP_FORBIDDEN);
+
         $departments = B3StorageInitiatorDepartment::query()
             ->orderBy('order_no')
             ->orderBy('id')
@@ -21,6 +25,8 @@ class B3StorageInitiatorDepartmentController extends Controller
 
     public function store(SaveB3StorageInitiatorDepartmentRequest $request): JsonResponse
     {
+        abort_unless($request->user()?->can('b3storage.master.manage'), Response::HTTP_FORBIDDEN);
+
         $department = B3StorageInitiatorDepartment::query()->create($request->validated());
 
         return response()->json([
@@ -29,8 +35,10 @@ class B3StorageInitiatorDepartmentController extends Controller
         ], 201);
     }
 
-    public function show(B3StorageInitiatorDepartment $initiatorDepartment): JsonResponse
+    public function show(Request $request, B3StorageInitiatorDepartment $initiatorDepartment): JsonResponse
     {
+        abort_unless($request->user()?->can('b3storage.master.view'), Response::HTTP_FORBIDDEN);
+
         return response()->json([
             'data' => $initiatorDepartment,
         ]);
@@ -40,6 +48,8 @@ class B3StorageInitiatorDepartmentController extends Controller
         SaveB3StorageInitiatorDepartmentRequest $request,
         B3StorageInitiatorDepartment $initiatorDepartment,
     ): JsonResponse {
+        abort_unless($request->user()?->can('b3storage.master.manage'), Response::HTTP_FORBIDDEN);
+
         $initiatorDepartment->update($request->validated());
 
         return response()->json([
@@ -48,8 +58,10 @@ class B3StorageInitiatorDepartmentController extends Controller
         ]);
     }
 
-    public function destroy(B3StorageInitiatorDepartment $initiatorDepartment): JsonResponse
+    public function destroy(Request $request, B3StorageInitiatorDepartment $initiatorDepartment): JsonResponse
     {
+        abort_unless($request->user()?->can('b3storage.master.manage'), Response::HTTP_FORBIDDEN);
+
         $initiatorDepartment->delete();
 
         return response()->json([
