+++ app/Http/Controllers/Api/B3Storage/B3StorageWasteTypeController.php
@@ -6,11 +6,15 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\B3Storage\SaveB3StorageWasteTypeRequest;
 use App\Models\B3Storage\B3StorageWasteType;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 
 class B3StorageWasteTypeController extends Controller
 {
-    public function index(): JsonResponse
+    public function index(Request $request): JsonResponse
     {
+        abort_unless($request->user()?->can('b3storage.master.view'), Response::HTTP_FORBIDDEN);
+
         $wasteTypes = B3StorageWasteType::query()
             ->orderBy('order_no')
             ->orderBy('id')
@@ -21,6 +25,8 @@ class B3StorageWasteTypeController extends Controller
 
     public function store(SaveB3StorageWasteTypeRequest $request): JsonResponse
     {
+        abort_unless($request->user()?->can('b3storage.master.manage'), Response::HTTP_FORBIDDEN);
+
         $wasteType = B3StorageWasteType::query()->create($request->validated());
 
         return response()->json([
@@ -29,8 +35,10 @@ class B3StorageWasteTypeController extends Controller
         ], 201);
     }
 
-    public function show(B3StorageWasteType $wasteType): JsonResponse
+    public function show(Request $request, B3StorageWasteType $wasteType): JsonResponse
     {
+        abort_unless($request->user()?->can('b3storage.master.view'), Response::HTTP_FORBIDDEN);
+
         return response()->json([
             'data' => $wasteType,
         ]);
@@ -40,6 +48,8 @@ class B3StorageWasteTypeController extends Controller
         SaveB3StorageWasteTypeRequest $request,
         B3StorageWasteType $wasteType,
     ): JsonResponse {
+        abort_unless($request->user()?->can('b3storage.master.manage'), Response::HTTP_FORBIDDEN);
+
         $wasteType->update($request->validated());
 
         return response()->json([
@@ -48,8 +58,10 @@ class B3StorageWasteTypeController extends Controller
         ]);
     }
 
-    public function destroy(B3StorageWasteType $wasteType): JsonResponse
+    public function destroy(Request $request, B3StorageWasteType $wasteType): JsonResponse
     {
+        abort_unless($request->user()?->can('b3storage.master.manage'), Response::HTTP_FORBIDDEN);
+
         $wasteType->delete();
 
         return response()->json([
